backend/requirements.txt
fastapi
uvicorn[standard]
beanie==1.25.0
motor==3.3.1
pymongo==4.6.3
pydantic-settings
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
email-validator
python-jose[cryptography]
python-multipart
httpx
resend
redis